backend/requirements.txt
# requirements.txt - Python 3.13 Compatible (AutoGen COMPLETELY REMOVED)

Django>=5.2
djangorestframework>=3.15.0
django-cors-headers>=4.3.1
pyjwt>=2.8.0
djangorestframework-simplejwt
aiohttp>=3.9.1
cryptography>=41.0.7
requests>=2.31.0
python-dotenv>=1.0.0

# PostgreSQL Database Dependencies
psycopg2-binary>=2.9.7
psycopg>=3.1.0

# Vector Search Dependencies (Python 3.13 compatible)
numpy>=1.24.0
pandas>=2.0.0
pymilvus>=2.3.4
sentence-transformers>=2.2.2
python-magic

# Public Chatbot Dependencies (isolated from main system)
chromadb==1.0.20
django-ratelimit>=4.1.0

# Django Milvus Search Dependencies
# (pymilvus and numpy already included above)
# django-redis>=5.0.0  # Optional, for caching

# Text Extraction Dependencies
PyPDF2>=3.0.1
python-docx>=0.8.11
openpyxl>=3.1.2
markdown>=3.5.1

# Gemini API for PDF and Image OCR
google-genai>=0.2.0
google-generativeai>=0.3.0
pdf2image>=1.16.0
PyMuPDF>=1.23.0

# OpenAI API for Content Summarization
openai>=1.0.0

# Performance monitoring and caching
psutil>=5.9.0
packaging>=21.0

# Agent Orchestration Dependencies (CUSTOM Implementation - NO AutoGen)
celery>=5.3.0
redis>=5.0.0
channels>=4.0.0
channels-redis>=4.1.0
asgiref>=3.7.0

# 🎯 CUSTOM AGENT ORCHESTRATION DEPENDENCIES (Our Own Implementation)
jsonschema>=4.0.0      # JSON Schema validation for workflow definitions
websockets>=11.0.0     # Real-time communication
anthropic>=0.3.0       # Anthropic Claude API

# 🎯 REAL-TIME EXECUTION DEPENDENCIES (Custom Implementation)
aiofiles>=23.0.0       # Async file operations
watchdog>=3.0.0        # File system monitoring
timeout-decorator>=0.5.0  # Function timeouts

# ❌ COMPLETELY REMOVED: All AutoGen dependencies
# pyautogen - REMOVED
# autogen-agentchat - REMOVED
# autogen-core - REMOVED
# docker - REMOVED (was only for AutoGen sandbox)
